# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис для мониторинга вакансий hh.ru.",
        "title": "HH Buff API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/rest",
    "paths": {
        "/data": {
            "get": {
                "description": "Возвращает срезы данных по ID запросов и диапазону дат",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data"
                ],
                "summary": "Получить исторические данные",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Дата начала (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата конца (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "multi",
                        "description": "Список ID запросов",
                        "name": "queries",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/hh_buff_internal_models.DBSnapshot"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/queries": {
            "get": {
                "description": "Возвращает список запросов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queries"
                ],
                "summary": "Получить список запросов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/hh_buff_internal_models.DBQuery"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queries"
                ],
                "summary": "Создать новый запрос",
                "parameters": [
                    {
                        "description": "Данные запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http.UploadQueryReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_http.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "hh_buff_internal_models.DBQuery": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "query": {
                    "$ref": "#/definitions/hh_buff_pkg_hh.GetVacanciesRequest"
                },
                "snapshots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/hh_buff_internal_models.DBSnapshot"
                    }
                }
            }
        },
        "hh_buff_internal_models.DBSnapshot": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "query": {
                    "$ref": "#/definitions/hh_buff_internal_models.DBQuery"
                },
                "query_id": {
                    "type": "integer"
                }
            }
        },
        "hh_buff_pkg_hh.GetVacanciesRequest": {
            "type": "object",
            "properties": {
                "area": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bottom_lat": {
                    "type": "number"
                },
                "currency": {
                    "type": "string"
                },
                "date_from": {
                    "type": "string"
                },
                "date_to": {
                    "type": "string"
                },
                "employer_id": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "employment": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "exclude_employer_id": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "experience": {
                    "type": "string"
                },
                "industry": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "label": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "left_lng": {
                    "type": "number"
                },
                "only_with_salary": {
                    "type": "boolean"
                },
                "order_by": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "period": {
                    "type": "integer"
                },
                "professional_role": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "right_lng": {
                    "type": "number"
                },
                "salary": {
                    "type": "integer"
                },
                "schedule": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "search_field": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "text": {
                    "type": "string"
                },
                "top_lat": {
                    "type": "number"
                }
            }
        },
        "internal_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_http.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_http.UploadQueryReq": {
            "type": "object",
            "required": [
                "name",
                "query"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "query": {
                    "$ref": "#/definitions/hh_buff_pkg_hh.GetVacanciesRequest"
                }
            }
        }
    }
}